rename work-efforts -> work-efforts-overview
