+++ alv-portal-ui/src/app/online-services/work-efforts/work-efforts-overview/modal-config.types.ts
