+++ alv-portal-ui/src/app/online-services/work-efforts/work-efforts-overview/proof-of-work-efforts/proof-of-work-efforts.component.html
