+++ alv-portal-ui/src/app/online-services/work-efforts/work-efforts-overview/work-effort/work-effort.component.ts
