+++ alv-portal-ui/src/app/online-services/work-efforts/work-efforts-overview/work-efforts-filter-modal/work-efforts-filter-modal.component.html
