+++ alv-portal-ui/src/app/online-services/work-efforts/work-efforts-overview/work-efforts-overview.component.html
